+++ src/tablecodec/codecs/builtins.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 from tablecodec.codecs._base import Codec
 from tablecodec.codecs.doctags import DocTagsTablesCodec
 from tablecodec.codecs.fintabnet import FinTabNetCodec
+from tablecodec.codecs.fintabnet_otsl import FinTabNetOTSLCodec
 from tablecodec.codecs.otsl import OTSL10Codec
 from tablecodec.codecs.pubtables1m import PubTables1MCodec
 from tablecodec.codecs.pubtabnet import PubTabNet10Codec, PubTabNet20Codec
@@ -30,5 +31,6 @@ BUILTIN_CODECS: tuple[Codec, ...] = (
     TableBankCodec(),
     PubTables1MCodec(),
     OTSL10Codec(),
+    FinTabNetOTSLCodec(),
     DocTagsTablesCodec(),
 )